test(agent): keep planner race guard within lint budgets
Problem:
The new planner concurrency regression exceeded existing repository test-file and task-file budgets.

Fix:
Move the concurrency test into its own focused file and trim redundant documentation while preserving the child-context boundary and behavior.

Verification:
- go test ./internal/agent -run TestCoordinatorSerializesConcurrentPlannerCalls -count=1
- go run ./tools/repolint
- git diff --check
