+++ internal/agent/coordinator_concurrency_test.go
@@ -1,0 +1,62 @@
+package agent
+
+import (
+	"context"
+	"sync"
+	"testing"
+
+	"reasonix/internal/event"
+	"reasonix/internal/provider"
+)
+
+type concurrentPlannerProvider struct {
+	mu    sync.Mutex
+	calls int
+}
+
+func (p *concurrentPlannerProvider) Name() string { return "concurrent-planner" }
+
+func (p *concurrentPlannerProvider) ContextBudgetPolicy() provider.ContextBudgetPolicy {
+	return provider.ContextBudgetPolicy{WindowMode: provider.ContextWindowIndependent}
+}
+
+func (p *concurrentPlannerProvider) Stream(_ context.Context, _ provider.Request) (<-chan provider.Chunk, error) {
+	p.mu.Lock()
+	p.calls++
+	p.mu.Unlock()
+	out := make(chan provider.Chunk, 2)
+	out <- provider.Chunk{Type: provider.ChunkText, Text: "plan"}
+	out <- provider.Chunk{Type: provider.ChunkDone}
+	close(out)
+	return out, nil
+}
+
+func TestCoordinatorSerializesConcurrentPlannerCalls(t *testing.T) {
+	planner := &concurrentPlannerProvider{}
+	coord := NewCoordinator(planner, NewSession("planner-sys"), nil, nil, Options{}, nil, 0, event.Discard, nil)
+
+	var wg sync.WaitGroup
+	errs := make(chan error, 2)
+	for _, input := range []string{"first", "second"} {
+		input := input
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+			_, err := coord.plan(context.Background(), input)
+			errs <- err
+		}()
+	}
+	wg.Wait()
+	close(errs)
+	for err := range errs {
+		if err != nil {
+			t.Fatalf("concurrent planner call: %v", err)
+		}
+	}
+	planner.mu.Lock()
+	calls := planner.calls
+	planner.mu.Unlock()
+	if calls != 2 {
+		t.Fatalf("planner calls = %d, want 2", calls)
+	}
+}
