+++ internal/agent/coordinator_test.go
@@ -7,7 +7,6 @@ import (
 	"reasonix/internal/event"
 	"slices"
 	"strings"
-	"sync"
 	"testing"
 
 	"reasonix/internal/provider"
@@ -86,58 +85,6 @@ func TestCoordinatorHandsPlanToExecutor(t *testing.T) {
 	// prefix grows prepend-only and stays cache-stable.
 	if n := len(plannerSess.Messages); n != 3 {
 		t.Errorf("planner session has %d messages, want 3", n)
-	}
-}
-
-type concurrentPlannerProvider struct {
-	mu    sync.Mutex
-	calls int
-}
-
-func (p *concurrentPlannerProvider) Name() string { return "concurrent-planner" }
-
-func (p *concurrentPlannerProvider) ContextBudgetPolicy() provider.ContextBudgetPolicy {
-	return provider.ContextBudgetPolicy{WindowMode: provider.ContextWindowIndependent}
-}
-
-func (p *concurrentPlannerProvider) Stream(_ context.Context, _ provider.Request) (<-chan provider.Chunk, error) {
-	p.mu.Lock()
-	p.calls++
-	p.mu.Unlock()
-	out := make(chan provider.Chunk, 2)
-	out <- provider.Chunk{Type: provider.ChunkText, Text: "plan"}
-	out <- provider.Chunk{Type: provider.ChunkDone}
-	close(out)
-	return out, nil
-}
-
-func TestCoordinatorSerializesConcurrentPlannerCalls(t *testing.T) {
-	planner := &concurrentPlannerProvider{}
-	coord := NewCoordinator(planner, NewSession("planner-sys"), nil, nil, Options{}, nil, 0, event.Discard, nil)
-
-	var wg sync.WaitGroup
-	errs := make(chan error, 2)
-	for _, input := range []string{"first", "second"} {
-		input := input
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			_, err := coord.plan(context.Background(), input)
-			errs <- err
-		}()
-	}
-	wg.Wait()
-	close(errs)
-	for err := range errs {
-		if err != nil {
-			t.Fatalf("concurrent planner call: %v", err)
-		}
-	}
-	planner.mu.Lock()
-	calls := planner.calls
-	planner.mu.Unlock()
-	if calls != 2 {
-		t.Fatalf("planner calls = %d, want 2", calls)
 	}
 }
